+++ NotificareKit/Sources/Internals/Modules/NotificareDeviceModuleImpl.swift
@@ -21,22 +21,28 @@ internal class NotificareDeviceModuleImpl: NSObject, NotificareModule, Notificar
 
     internal func configure() {
         // Listen to timezone changes
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(updateDeviceTimezone),
-                                               name: UIApplication.significantTimeChangeNotification,
-                                               object: nil)
+        NotificationCenter.default.upsertObserver(
+            self,
+            selector: #selector(updateDeviceTimezone),
+            name: UIApplication.significantTimeChangeNotification,
+            object: nil
+        )
 
         // Listen to language changes
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(updateDeviceLanguage),
-                                               name: NSLocale.currentLocaleDidChangeNotification,
-                                               object: nil)
+        NotificationCenter.default.upsertObserver(
+            self,
+            selector: #selector(updateDeviceLanguage),
+            name: NSLocale.currentLocaleDidChangeNotification,
+            object: nil
+        )
 
         // Listen to 'background refresh status' changes
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(updateDeviceBackgroundAppRefresh),
-                                               name: UIApplication.backgroundRefreshStatusDidChangeNotification,
-                                               object: nil)
+        NotificationCenter.default.upsertObserver(
+            self,
+            selector: #selector(updateDeviceBackgroundAppRefresh),
+            name: UIApplication.backgroundRefreshStatusDidChangeNotification,
+            object: nil
+        )
     }
 
     internal func launch() async throws {
@@ -421,7 +427,7 @@ internal class NotificareDeviceModuleImpl: NSObject, NotificareModule, Notificar
         return userData
     }
 
-    public func updateUserData(_ userData: NotificareUserData, _ completion: @escaping NotificareCallback<Void>) {
+    public func updateUserData(_ userData: [String: String?], _ completion: @escaping NotificareCallback<Void>) {
         Task {
             do {
                 try await updateUserData(userData)
@@ -432,7 +438,7 @@ internal class NotificareDeviceModuleImpl: NSObject, NotificareModule, Notificar
         }
     }
 
-    public func updateUserData(_ userData: NotificareUserData) async throws {
+    public func updateUserData(_ userData: [String: String?]) async throws {
         guard Notificare.shared.isReady, let device = storedDevice else {
             throw NotificareError.notReady
         }
@@ -446,7 +452,7 @@ internal class NotificareDeviceModuleImpl: NSObject, NotificareModule, Notificar
             .response()
 
         // Update current device properties.
-        storedDevice?.userData = userData
+        storedDevice?.userData = userData.compactMapValues { $0 }
     }
 
     // MARK: - Internal API