+++ NotificarePushUIKit/Sources/Action Handlers/NotificareAppActionHandler.swift
@@ -19,6 +19,8 @@ public class NotificareAppActionHandler: NotificareBaseActionHandler {
                     Task {
                         try? await Notificare.shared.createNotificationReply(notification: self.notification, action: self.action)
                     }
+
+                    self.dismiss()
                 }
             }
         } else {