+++ NotificarePushUIKit/Sources/Action Handlers/NotificareMailActionHandler.swift
@@ -7,14 +7,6 @@ import NotificareKit
 import NotificareUtilitiesKit
 
 public class NotificareMailActionHandler: NotificareBaseActionHandler {
-    private let sourceViewController: UIViewController
-
-    internal init(notification: NotificareNotification, action: NotificareNotification.Action, sourceViewController: UIViewController) {
-        self.sourceViewController = sourceViewController
-
-        super.init(notification: notification, action: action)
-    }
-
     internal override func execute() {
         guard let target = action.target, MFMailComposeViewController.canSendMail() else {
             DispatchQueue.main.async {
@@ -33,20 +25,6 @@ public class NotificareMailActionHandler: NotificareBaseActionHandler {
         composer.setMessageBody(NotificareLocalizable.string(resource: .actionMailBody), isHTML: false)
 
         sourceViewController.presentOrPush(composer)
-    }
-
-    private func dismiss() {
-        if let rootViewController = UIApplication.shared.rootViewController, rootViewController.presentedViewController != nil {
-            rootViewController.dismiss(animated: true, completion: nil)
-        } else {
-            if sourceViewController is UIAlertController {
-                UIApplication.shared.rootViewController?.dismiss(animated: true, completion: nil)
-            } else {
-                sourceViewController.dismiss(animated: true) {
-                    self.sourceViewController.becomeFirstResponder()
-                }
-            }
-        }
     }
 }