+++ Sample/Sample/AppDelegate.swift
@@ -127,16 +127,11 @@ extension AppDelegate: NotificarePushDelegate {
     }
 
     internal func notificare(_ notificarePush: any NotificarePush, didChangeSubscription subscription: NotificarePushSubscription?) {
-        Logger.main.info("Notificare: subscription changed: \(String(describing: subscription))")
+        Logger.main.info("Notification center subscription changed: \(String(describing: subscription))")
     }
 
     internal func notificare(_: NotificarePush, didChangeNotificationSettings allowedUI: Bool) {
-        Logger.main.info("Notificare: notification settings changed: \(allowedUI)")
-
-        NotificationCenter.default.post(
-            name: .notificationSettingsChanged,
-            object: nil
-        )
+        Logger.main.info("Notification center allowedUI changed: \(allowedUI)")
     }
 
     internal func notificare(_: NotificarePush, didReceiveSystemNotification notification: NotificareSystemNotification) {
@@ -228,23 +223,11 @@ extension AppDelegate: NotificarePushUIDelegate {
 
 extension AppDelegate: NotificareInboxDelegate {
     internal func notificare(_: NotificareInbox, didUpdateInbox items: [NotificareInboxItem]) {
-        Logger.main.info("Inbox has loaded. Total = \(items.count)")
-
-        NotificationCenter.default.post(
-            name: .inboxUpdated,
-            object: nil,
-            userInfo: ["items": items]
-        )
+        Logger.main.info("Delegate inbox update. Total = \(items.count)")
     }
 
     internal func notificare(_: NotificareInbox, didUpdateBadge badge: Int) {
-        Logger.main.info("Badge update. Unread = \(badge)")
-
-        NotificationCenter.default.post(
-            name: .badgeUpdated,
-            object: nil,
-            userInfo: ["badge": badge]
-        )
+        Logger.main.info("Delegate badge update. Unread = \(badge)")
     }
 }