+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## 4.1.0
+
+- Introduce `itemsStream` and `badgeStream` Combine publishers for inbox module
+- Introduce `subscriptionStream` and `allowedUIStream` Combine publishers for push module
+- Fix issue where the notification UI was not dismissed when some actions were executed
+- Allow unsetting user data fields
+- Run the configuration flow on the main thread
+- Fix callback with keyboard safe area insets
+- Improve CoreData threading
+- Change `NotificareRegion.timeZoneOffset` attribute from `int` to `float` to support half-hour timezones
+
 ## 4.0.1
 
 - Add documentation to public methods.